+++ app/Http/Controllers/Dashboard/DepartmentController.php
@@ -47,7 +47,9 @@ class DepartmentController extends DashboardController
         ];
         $breadcrumbs = $this->breadcrumbs;
         $title = 'BBPP Batangkaluku - Edit Bagian / Bidang';
-        $data = Department::find($id);
+        if (!$data = Department::find($id)){
+            return abort(404,'Not Found');
+        }
         $action = route('dashboard.bagian.edit',$id);
         $method = 'post';
         $redirect = route('dashboard.bagian');
@@ -85,7 +87,9 @@ class DepartmentController extends DashboardController
         ]);
         try {
             DB::beginTransaction();
-            $model = Department::find($id);
+            if (!$model = Department::find($id)){
+                return abort(404,'Not Found');
+            }
             $model->update($data);
         } catch (\Exception $exception){
             DB::rollBack();
@@ -100,7 +104,9 @@ class DepartmentController extends DashboardController
     }
 
     public function delete($id){
-        $model = Department::find($id);
+        if (!$model = Department::find($id)){
+            return abort(404,'Not Found');
+        }
         try {
             $model->delete();
         }catch (\Exception $exception){
